requirements.txt
streamlit
torch==1.6.0+cu92 
torchvision==0.7.0+cu92 
fastai==1.0.61
pillow==6.2.1

